refactor: reuse invitation contact item
